a.kashirina/backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-RESTX==0.5.1
Flask-JWT-Extended==4.3.1
Flask-Bcrypt==0.7.1
Flask-CORS==3.0.10
psycopg2-binary==2.9.1
python-dotenv==0.19.0
gunicorn==20.1.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.8.0
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
python-dateutil==2.8.2
PyJWT==2.8.0
Werkzeug==2.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
SQLAlchemy==1.4.23
alembic==1.13.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
requests==2.31.0
urllib3==2.2.0
# Дополнительные библиотеки для разработки
pytest-flask==1.3.0
coverage==7.4.1
factory-boy==3.3.0
Faker==22.6.0
python-jose==3.3.0
email-validator==2.1.0.post1
debugpy==1.6.7 